Modify tfe client mocks to meet some new requirements
1. Hook up `MockPlans.ReadJSONOutput` to test fixtures, when present.
This method has been implemented for ages, and has had a backing store for
unredacted plan json, but has been effectively a no-op since nothing ever fills
that backing store. So, when creating a mock plan, make an attempt to read
unredacted json and stow it in the mocks on success.

2. Make it possible to get the entire MockClient for a test backend
In order to test some things, I'm going to need to mess with the internal state
of runs and plans beyond what the go-tfe client API allows. I could add magic
special-casing to the mock API methods, or I could locate the shenanigans next
to the test that actually exploits it. The latter seems more comprehensible, but
I need access to the full mock client struct in order to mess with its interior.

3. Fill in some missing expectations around HasChanges when retrieving a run +
plan.