Add new plans.Quality type to represent display-relevant facts about a plan
We need to be able to exchange the facts represented by
`jsonformat.PlanRendererOpt` across some package boundaries, but that type's
package is implicated in too many dependency chains to be safe for that purpose.
So, let's make a new one. The plans package seems safe to import from all the
places that must emit or accept this info.